made the page more mobile friendly by changing what is displayed by the outputTable function, added in confirmation for delete message, other styling changes

+++ update-team.php
@@ -52,14 +52,9 @@ $query=mysqli_query($db, $res);
 function outputTable($query){
     while($row=mysqli_fetch_array($query)){
         echo "<tr>";
-        echo  "<td>"; echo $row["teamName"]."</td>";
-        echo  "<td>"; echo $row["coachFirstName"]."</td>";
-        echo  "<td>"; echo $row["coachLastName"]."</td>";
-        echo  "<td>"; echo $row["coachEmail"]."</td>";
-        echo  "<td>"; echo $row["ageGroup"]."u</td>";
-        echo  "<td>"; echo $row["teamLocation"]."</td>";
+        echo  "<td>"; echo $row["teamIdentifier"]."</td>";
         echo  "<td>"; ?> <a href="edit-team.php?id=<?php echo $row["teamId"];?>"> <button type="button" class= "btn btn-success">Edit</button></a> <?php echo "</td>"; #update team
-        echo  "<td>"; ?> <a href="delete-team.php?id=<?php echo $row["teamId"];?>"> <button type="button" class= "btn btn-danger">Delete</button></a> <?php echo "</td>"; #delete team
+        echo  "<td>"; ?> <a href="delete-team.php?id=<?php echo $row["teamId"] ?> "onclick="return confirm('Are you sure you want to delete <?php echo $row['teamIdentifier'] ?>?') "><button type="button" class= "btn btn-danger">Delete</button></a> <?php echo "</td>"; #delete team
         echo"</tr>";
       }
 }
@@ -80,7 +75,9 @@ if (isset($_POST['showAll'])){
 
 <!-- START OF BODY -->
 <body>
-<div class="text-center p-2 mt-3" >
+
+<h1 class='centerContent mt-4'>Teams</h1>
+<div class="text-center p-2 mt-2" >
 <form name="search_form" method="POST" action="update-team.php">
 Search: <input type="text" name="search_box" value="" />
 
@@ -92,17 +89,12 @@ Search: <input type="text" name="search_box" value="" />
 <div class="col-lg-12 p-2">
 <?=$promptMessage()
 ?>
-<table class="table table-bordered">
+<table class="table table-bordered mx-lg-2 centerContent text-center">
 <tbody>
     <thead>
       <tr>
         <th>Team Name</th>
-        <th>Coach First Name</th>
-        <th>Coach Last Name</th>
-        <th>Coach Email</th>
-        <th>Age Group</th>
-        <th>Team Location</th>
-        <th>Update</th>
+        <th>Edit</th>
         <th>Delete</th>
       </tr>
       <?=outputTable($query)?>
